Version 0.3
Files are now actually downloaded and saved!
Separated projects sub and the sub that actually has the file downloads on.
Files are saved in a folder specific to the version and site selected.
Uses api.cfwidget.com because curse cant be bothered to document their own api.
There are definitely crashes/bugs and other random behaviour, but it should at least download something before crashing.